ncurses 5.9 - patch 20130112
+ correct prototype in manpage for vid_puts.
+ drop ncurses/tty/tty_display.h, ncurses/tty/tty_input.h, since they
  are unused in the current driver model.
+ modify mvcur to use stdout except when called within the ncurses
  library.
+ modify vidattr and vid_attr to use stdout as documented in manpage.
+ amend changes made to buffering in 20120825 so that the low-level
  putp() call uses stdout rather than ncurses' internal buffering.
  The putp_sp() call does the same, for consistency (Redhat #892674).

+++ test/view.c
@@ -50,7 +50,7 @@
  * scroll operation worked, and the refresh() code only had to do a
  * partial repaint.
  *
- * $Id: view.c,v 1.90 2013/01/05 23:18:13 tom Exp $
+ * $Id: view.c,v 1.91 2013/01/13 01:00:44 tom Exp $
  */
 
 #include <test.priv.h>
@@ -274,7 +274,7 @@ main(int argc, char *argv[])
 	case 'T':
 	    {
 		char *next = 0;
-		int tvalue = strtol(optarg, &next, 0);
+		int tvalue = (int) strtol(optarg, &next, 0);
 		if (tvalue < 0 || (next != 0 && *next != 0))
 		    usage();
 		trace((unsigned) tvalue);
